Module Alternative1.Extend

Functor building an instance of Alternative given a Alternative.Base and Applicative implementation.

Parameters

Instance : Alternative1

Applicative : Applicative1 with type 'a t := 'a Instance.t

Signature

type 'a t

val empty : 'a t

empty is the identity of <|>

val (<|>) : 'a t -> 'a t -> 'a t

a <|> b in associative binary operation of a and b.

val some : 'a t -> 'a list t

some a is one or more occurrences of a.

val many : 'a t -> 'a list t

many a is zero or more occurrences of a.