Module Alternative2.Extend

Functor building an instance of Alternative2 given a Alternative2.Base and Applicative2 implementation.

Parameters

Instance : Alternative2

Applicative : Applicative2 with type ('a, 'x) t := ('a, 'x) Instance.t

Signature

type ('a, 'x) t

val empty : ('a, 'x) t

empty is the identity of <|>

val (<|>) : ('a, 'x) t -> ('a, 'x) t -> ('a, 'x) t

a <|> b in associative binary operation of a and b.

val some : ('a, 'x) t -> ('a list, 'x) t

some a is one or more occurrences of a.

val many : ('a, 'x) t -> ('a list, 'x) t

many a is zero or more occurrences of a.