Module type Alternative2.Extension

type ('a, 'x) t

val empty : ('a, 'x) t

empty is the identity of <|>

val (<|>) : ('a, 'x) t -> ('a, 'x) t -> ('a, 'x) t

a <|> b in associative binary operation of a and b.

val some : ('a, 'x) t -> ('a list, 'x) t

some a is one or more occurrences of a.

val many : ('a, 'x) t -> ('a list, 'x) t

many a is zero or more occurrences of a.