Parameter To_functor.1-Monad

type ('a, 'x) t

The type of the monadic values.

val return : 'a -> ('a, 'x) t

pure x injects the value x into the monadic type.

val bind : ('a -> ('b, 'x) t) -> ('a, 'x) t -> ('b, 'x) t

bind f m sequentially composes two actions, passing any value produced by m as an argument to f.