Module type Monad2.Extension

type ('a, 'x) t

The type of the monadic values.

val return : 'a -> ('a, 'x) t

return x injects the value x into the monadic type.

val bind : ('a -> ('b, 'x) t) -> ('a, 'x) t -> ('b, 'x) t

val (>>=) : ('a, 'x) t -> ('a -> ('b, 'x) t) -> ('b, 'x) t

val (=<<) : ('a -> ('b, 'x) t) -> ('a, 'x) t -> ('b, 'x) t

Sequentially compose two actions, passing any value produced by the first as an argument to the second.

>>= and =<< are the infix versions of bind.

val (>>) : ('a, 'x) t -> ('b, 'x) t Stdlib.Lazy.t -> ('b, 'x) t

m1 >> fun () -> m2 sequentially composes two actions, discarding any value produced by m1, like sequencing operators (such as the semicolon) in imperative languages.

val sequence : ('a, 'x) t list -> ('a list, 'x) t

sequence l evaluates each monadic action in l from left to right, and collects the results.

val join : (('a, 'x) t, 'x) t -> ('a, 'x) t

join m removes one level of monadic structure, projecting its bound argument into the outer level.