Module type Control.Monad

type 'a t

The type of the monadic values.

val return : 'a -> 'a t

return x injects a value x into a monadic value.

val bind : ('a -> 'b t) -> 'a t -> 'b t

bind f m sequentially composes two computations, passing any value produced by m as an argument to f.