Module type Control.State

type state

type 'a t = state -> 'a * state

include Monad1.Extension with type 'a t := 'a t

type 'a t

The type of the monadic values.

val return : 'a -> 'a t

return x injects the value x into the monadic type.

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (=<<) : ('a -> 'b t) -> 'a t -> 'b t

val bind : ('a -> 'b t) -> 'a t -> 'b t

Sequentially compose two actions, passing any value produced by the first as an argument to the second.

>>= and =<< are the infix versions of bind.

val (>>) : 'a t -> (unit -> 'b t) -> 'b t

m1 >> fun () -> m2 sequentially composes two actions, discarding any value produced by m1, like sequencing operators (such as the semicolon) in imperative languages.

val sequence : 'a t list -> 'a list t

sequence l evaluates each monadic action in l from left to right, and collects the results.

val join : 'a t t -> 'a t

join m removes one level of monadic structure, projecting its bound argument into the outer level.

val get : state t

val put : state -> unit t

val run : 'a t -> state -> 'a * state